Bets error handling

+++ backend/queries/bets/bets.js
@@ -1,13 +1,31 @@
 const db = require("./../../db/db")
 
+const isBetExisting = async (betId) => {
+    try {
+        let bet = await db.any("SELECT * FROM bets WHERE id=$1", betId);
+        if(bet.length) {
+            return true;
+        } else {
+            return false;
+        }
+    } catch(err) {
+        console.log(err);
+    }
+}
+
 const getBets = async (req, res, next) => {
     try {
         let bets = await db.any("SELECT * FROM bets INNER JOIN users ON bets.better_id=users.id");
-        res.status(200).json({
-            bets,
-            status: "success",
-            message: "Grabbed all bets"
-        })
+        if(bets.length) {
+            res.status(200).json({
+                bets,
+                status: "success",
+                message: "Grabbed all bets"
+            })
+        } else {
+            throw {status: 404, error: "No bets found"}
+        }
+        
     } catch (err) {
         next(err)
     }
@@ -16,12 +34,17 @@ const getBets = async (req, res, next) => {
 const getBetsById = async (req,res,next) => {
     let {betId} = req.params
     try{ 
-        let bet = await db.any("SELECT from bets WHERE id=$1 INNER JOIN users ON bets.better_id=users.id", betId);
-        res.status(200).json({
-            bet,
-            status: "sucess",
-            message: "Grabbed bet by ID"
-        })
+        if(await isBetExisting(betId)) {
+            let bet = await db.any("SELECT from bets WHERE id=$1 INNER JOIN users ON bets.better_id=users.id", betId);
+            res.status(200).json({
+                bet,
+                status: "sucess",
+                message: "Grabbed bet by ID"
+            })
+        } else {
+            throw {status: 404, error: "Bet doesn't exist"}
+        }
+        
     } catch(err) {
         next(err)
     }
@@ -30,11 +53,16 @@ const getBetsById = async (req,res,next) => {
 const getBetsNoTaker = async (req,res,next) => {
     try {
         let bets = await db.any("SELECT * FROM bets INNER JOIN users ON bets.better_id=users.id WHERE taker_id IS NULL");
-        res.status(200).json({
-            bets, 
-            status: "sucess",
-            message: "Grabbed bets with no taker"
-        })
+        if(bets.length) {
+            res.status(200).json({
+                bets, 
+                status: "sucess",
+                message: "Grabbed bets with no taker"
+            })
+        } else {
+            throw {status: 404, error: "No bets found"}
+        }
+        
     } catch(err){
         next(err)
     }
@@ -56,12 +84,18 @@ const postBets = async (req,res,next) => {
 const patchBets = async (req,res,next) => {
     try{
         let {taker_id} = req.body;
-        let updateBet = await db.one("UPDATE bets SET taker_id=$1 RETURNING *", taker_id);
-        res.status(200).json({
-            updateBet,
-            status:"Sucess",
-            message: "Updated bet"
-        })
+        let {betId} = req.params;
+        if(await isBetExisting(betId)) {
+            let updateBet = await db.one("UPDATE bets SET taker_id=$1 WHERE id=$2 RETURNING *", [taker_id, betId]);
+            res.status(200).json({
+                updateBet,
+                status:"Sucess",
+                message: "Updated bet"
+            })
+        } else {
+            throw {status: 404, error: "Bet doesn't exist"}
+        }
+        
     } catch(err) {
         next(err)
     }
@@ -70,13 +104,16 @@ const patchBets = async (req,res,next) => {
 const deleteBet = async (req,res,next) => {
     try{
         let {betId} = req.params;
-        let eraseBet = await db.one("DELETE FROM bets WHERE id=$1 RETURNING *", betId);
-        res.status(200).json({
-            eraseBet,
-            status: "Sucess",
-            message: "Deleted Bet"
-        })
-
+        if(await isBetExisting(betId)) {
+            let eraseBet = await db.one("DELETE FROM bets WHERE id=$1 RETURNING *", betId);
+            res.status(200).json({
+                eraseBet,
+                status: "Sucess",
+                message: "Deleted Bet"
+            })
+        } else {
+            throw {status: 404, error: "Bet doesn't exist"}
+        }
     } catch(err){
         next(err)
     }
